#34 Add Data Compatibility Checker
I have added a Python script (data_compatibility_checker.py) and used it as a module in the FPAR_Aggregation notebook to give the users a function to check the size identically between crop mask and raster data. Also, check the coordinate reference system (CRS) for all data, including the shapefile.